refactor(Movies and SavedMovies):
- move to pages folder.

meta: Level-2
