+++ src/containers/index.js
@@ -1,8 +1,6 @@
 export { LoginContainer } from './LoginContainer';
-export { Movies } from './Movies';
 export { NavigationContainer } from './NavigationContainer';
 export { ProfileContainer } from './ProfileContainer';
 export { RegisterContainer } from './RegisterContainer';
-export { SavedMovies } from './SavedMovies';
 export { HeaderContainer } from './HeaderContainer';
 export { ErrorPageContainer } from './ErrorPageContainer';
