+++ src/pages/Movies/styled.jsx
