+++ src/pages/index.js
@@ -1,0 +1,2 @@
+export { SavedMovies } from './SavedMovies';
+export { Movies } from './Movies';
